+++ lib/l10n/app_localizations_cs.dart
@@ -3692,6 +3692,9 @@ class AppLocalizationsCs extends AppLocalizations {
   String get devicesTitle => 'Připojená zařízení';
 
   @override
+  String get devicesTab => 'Zařízení';
+
+  @override
   String get devicesAll => 'Vše';
 
   @override