+++ lib/l10n/app_localizations_ru.dart
@@ -3741,6 +3741,9 @@ class AppLocalizationsRu extends AppLocalizations {
   String get devicesTitle => 'Подключённые устройства';
 
   @override
+  String get devicesTab => 'Устройства';
+
+  @override
   String get devicesAll => 'Все';
 
   @override